fix: mark slots as STOPPED instead of voiding when drug is stopped without administration
When a drug order is stopped and no medication has been administered yet,
the slots were being voided (hidden from all queries). This caused stopped
drugs to disappear from Nursing Tasks, Drug Chart, and IPD Medication
section. Changed the else branch in handleDrugOrderStop() to set slot
status to STOPPED instead of voiding.

- Remove unused logger imports from IPDScheduleServiceImpl
- Add unit tests for handleDrugOrderStop (3 scenarios: no admin, partial admin, GP disabled)
