+++ omod/src/test/java/org/openmrs/module/ipd/web/service/impl/IPDScheduleServiceImplTest.java
@@ -5,15 +5,18 @@ import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
-import org.mockito.junit.MockitoJUnitRunner;
 import org.openmrs.Concept;
 import org.openmrs.DrugOrder;
 import org.openmrs.Patient;
 import org.openmrs.Visit;
+import org.openmrs.api.AdministrationService;
 import org.openmrs.api.ConceptService;
 import org.openmrs.api.OrderService;
 import org.openmrs.api.PatientService;
 import org.openmrs.api.VisitService;
+import org.openmrs.api.context.Context;
+import org.openmrs.module.emrapi.encounter.domain.EncounterTransaction;
+import org.openmrs.module.ipd.api.events.IPDEventManager;
 import org.openmrs.module.ipd.api.model.MedicationAdministration;
 import org.openmrs.module.ipd.api.model.Reference;
 import org.openmrs.module.ipd.api.model.Schedule;
@@ -25,10 +28,14 @@ import org.openmrs.module.ipd.api.service.SlotService;
 import org.openmrs.module.ipd.web.contract.ScheduleMedicationRequest;
 import org.openmrs.module.ipd.web.factory.SlotFactory;
 import org.openmrs.module.ipd.web.model.SlotTimeCreationResult;
+import org.powermock.api.mockito.PowerMockito;
+import org.powermock.core.classloader.annotations.PrepareForTest;
+import org.powermock.modules.junit4.PowerMockRunner;
 
 import java.time.LocalDateTime;
 import java.util.Arrays;
 import java.util.Collections;
+import java.util.Date;
 import java.util.Optional;
 
 import org.openmrs.api.APIException;
@@ -36,9 +43,11 @@ import org.openmrs.api.APIException;
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.fail;
 import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.Mockito.*;
 
-@RunWith(MockitoJUnitRunner.class)
+@RunWith(PowerMockRunner.class)
+@PrepareForTest({Context.class})
 public class IPDScheduleServiceImplTest {
 
     @InjectMocks
@@ -53,6 +62,8 @@ public class IPDScheduleServiceImplTest {
     @Mock private PatientService patientService;
     @Mock private OrderService orderService;
     @Mock private SlotTimeCreationService slotTimeCreationService;
+    @Mock private AdministrationService administrationService;
+    @Mock private IPDEventManager eventManager;
 
     private Patient patient;
     private Visit visit;
@@ -86,6 +97,9 @@ public class IPDScheduleServiceImplTest {
                 .thenReturn(medicationRequestConcept);
         when(referenceService.getReferenceByTypeAndTargetUUID(Patient.class.getTypeName(), "patient-uuid"))
                 .thenReturn(Optional.of(patientReference));
+
+        PowerMockito.mockStatic(Context.class);
+        when(Context.getAdministrationService()).thenReturn(administrationService);
         when(slotTimeCreationService.createSlotsStartTimeFrom(any(), any()))
                 .thenReturn(SlotTimeCreationResult.withoutCrossingTags(Collections.emptyList()));
     }
@@ -418,5 +432,98 @@ public class IPDScheduleServiceImplTest {
         assertNotNull("Slots should not be null", result);
         assert(result.size() == 1);
         verify(slotService).getSlotsBySubjectReferenceIncludingAdministeredTimeFrame(patientReference, startDate, endDate, visit);
+    }
+
+    private EncounterTransaction.DrugOrder buildStoppedDrugOrder(String previousOrderUuid, Date dateStopped) {
+        EncounterTransaction.DrugOrder drugOrder = new EncounterTransaction.DrugOrder();
+        drugOrder.setPreviousOrderUuid(previousOrderUuid);
+        drugOrder.setDateStopped(dateStopped);
+        return drugOrder;
+    }
+
+    private Slot buildSlot(String uuid, Slot.SlotStatus status, MedicationAdministration administration, LocalDateTime startDateTime) {
+        Slot slot = new Slot();
+        slot.setUuid(uuid);
+        slot.setStatus(status);
+        slot.setMedicationAdministration(administration);
+        slot.setStartDateTime(startDateTime);
+        return slot;
+    }
+
+    private Date toDate(LocalDateTime ldt) {
+        return java.sql.Timestamp.valueOf(ldt);
+    }
+
+    @Test
+    public void shouldMarkAllSlotsStopped_WhenNoSlotsAdministered() {
+        when(administrationService.getGlobalProperty("bahmni-ipd.allowSlotStopOnDrugOrderStop", "false")).thenReturn("true");
+
+        LocalDateTime now = LocalDateTime.now();
+        Date dateStopped = toDate(now.minusHours(1));
+
+        EncounterTransaction encounterTransaction = new EncounterTransaction();
+        encounterTransaction.setPatientUuid("patient-uuid");
+        encounterTransaction.setDrugOrders(Collections.singletonList(
+                buildStoppedDrugOrder("prev-order-uuid", dateStopped)));
+
+        Slot slot1 = buildSlot("slot-1", Slot.SlotStatus.SCHEDULED, null, now.plusHours(1));
+        Slot slot2 = buildSlot("slot-2", Slot.SlotStatus.SCHEDULED, null, now.plusHours(3));
+
+        when(slotService.getSlotsBySubjectReferenceIdAndServiceTypeAndOrderUuids(
+                patientReference, medicationRequestConcept, Collections.singletonList("prev-order-uuid")))
+                .thenReturn(Arrays.asList(slot1, slot2));
+
+        service.handlePostProcessEncounterTransaction(null, encounterTransaction);
+
+        verify(slotService, times(2)).saveSlot(argThat(slot -> slot.getStatus() == Slot.SlotStatus.STOPPED));
+    }
+
+    @Test
+    public void shouldMarkNonAdministeredSlotsStopped_WhenSomeSlotsAlreadyAdministered() {
+        when(administrationService.getGlobalProperty("bahmni-ipd.allowSlotStopOnDrugOrderStop", "false")).thenReturn("true");
+
+        LocalDateTime now = LocalDateTime.now();
+        Date dateStopped = toDate(now.minusHours(2));
+
+        EncounterTransaction encounterTransaction = new EncounterTransaction();
+        encounterTransaction.setPatientUuid("patient-uuid");
+        encounterTransaction.setDrugOrders(Collections.singletonList(
+                buildStoppedDrugOrder("prev-order-uuid", dateStopped)));
+
+        Slot administered = buildSlot("slot-1", Slot.SlotStatus.COMPLETED,
+                new MedicationAdministration(), now.minusHours(6));
+
+        Slot notAdministeredFuture = buildSlot("slot-2", Slot.SlotStatus.SCHEDULED,
+                null, now.plusHours(1));
+
+        Slot notAdministeredPast = buildSlot("slot-3", Slot.SlotStatus.SCHEDULED,
+                null, now.minusHours(4));
+
+        when(slotService.getSlotsBySubjectReferenceIdAndServiceTypeAndOrderUuids(
+                patientReference, medicationRequestConcept, Collections.singletonList("prev-order-uuid")))
+                .thenReturn(Arrays.asList(administered, notAdministeredFuture, notAdministeredPast));
+
+        service.handlePostProcessEncounterTransaction(null, encounterTransaction);
+
+        verify(slotService, times(1)).saveSlot(argThat(slot ->
+                slot.getUuid().equals("slot-2") && slot.getStatus() == Slot.SlotStatus.STOPPED));
+
+        verify(slotService, never()).saveSlot(argThat(slot -> slot.getUuid().equals("slot-1")));
+        verify(slotService, never()).saveSlot(argThat(slot -> slot.getUuid().equals("slot-3")));
+    }
+
+    @Test
+    public void shouldNotStopSlots_WhenGlobalPropertyIsFalse() {
+        when(administrationService.getGlobalProperty("bahmni-ipd.allowSlotStopOnDrugOrderStop", "false")).thenReturn("false");
+
+        EncounterTransaction encounterTransaction = new EncounterTransaction();
+        encounterTransaction.setPatientUuid("patient-uuid");
+        encounterTransaction.setDrugOrders(Collections.singletonList(
+                buildStoppedDrugOrder("prev-order-uuid", new Date())));
+
+        service.handlePostProcessEncounterTransaction(null, encounterTransaction);
+
+        verify(referenceService, never()).getReferenceByTypeAndTargetUUID(anyString(), anyString());
+        verify(slotService, never()).saveSlot(any());
     }
 }
